requirements.txt
loguru==0.7.2
python-dotenv==1.0.0
python-telegram-bot==21.1.1
